requirements.txt
# SemantiQ: Bloom's Taxonomy Classifier

## Core ML
torch>=2.0.0
transformers>=4.35.0
scikit-learn>=1.3.0

## Word Embeddings
fasttext>=0.9.2
gensim>=4.3.0

## Data Processing
pandas>=2.0.0
numpy>=1.24.0
nltk>=3.8.0
spacy>=3.0.0
# Note: After install, run: python -m spacy download en_core_web_md

## Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

## Utilities
pyyaml>=6.0
tqdm>=4.65.0
jupyter>=1.0.0

## Testing
pytest>=7.4.0
